# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Safe2Gether API
  version: 1.0.0
  description: |
    REST API for Safe2Gether — a community safety app for the DACH region.

    **Core principle**: Safety is always free. Premium is comfort and community.

    All timestamps are ISO 8601 (UTC). All IDs are UUID v4.

    WebSocket endpoint `/ws/alarm/{user_id}` is not covered by this spec —
    OpenAPI 3.1 does not natively support WebSocket. See `x-websocket` at the
    bottom of this document; a full AsyncAPI spec is planned for v2.0.
  license:
    name: AGPL v3
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    url: https://github.com/StekkenRyan
servers:
  - url: https://api.safe2gether.de
    description: Production (Cloudflare Tunnel → sora)
  - url: http://localhost:5000
    description: Local development (Docker Compose)
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT (1 h) obtained from POST /api/v1/auth/signin

  # ── Schemas ──────────────────────────────────────────────
  schemas:
    User:
      type: object
      required: [id, auth_provider, reputation_score, reputation_level, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: Present for google and email providers
        phone_number:
          type: string
          example: "+4915112345678"
        phone_verified:
          type: boolean
          default: false
        auth_provider:
          type: string
          enum: [apple, google, email]
        reputation_score:
          type: integer
          minimum: 0
          example: 75
        reputation_level:
          type: string
          enum: [normal, low, very_low]
          description: |
            - **normal**: full access
            - **low**: can send alarms; does not receive community alerts;
              false-alarm notice shown to recipients
            - **very_low**: manual review queue; alarm sending always remains possible
              (safety is never gated)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SignInApple:
      type: object
      required: [provider, id_token]
      properties:
        provider:
          type: string
          enum: [apple]
        id_token:
          type: string
          description: Identity token from Apple Sign In
    SignInGoogle:
      type: object
      required: [provider, id_token]
      properties:
        provider:
          type: string
          enum: [google]
        id_token:
          type: string
          description: ID token from Google Sign In
    SignInEmail:
      type: object
      required: [provider, email, password]
      properties:
        provider:
          type: string
          enum: [email]
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    AuthResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
          description: Short-lived JWT (1 hour)
        refresh_token:
          type: string
          description: Long-lived refresh token (30 days)
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/User'

    # ── Standard error ────────────────────────────────────────
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          example: "Unauthorized"
        code:
          type: string
          example: "TOKEN_EXPIRED"
        detail:
          type: string
          example: "The access token has expired. Use /auth/refresh to obtain a new one."

  # ── Reusable responses ────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ══════════════════════════════════════════════════════════
# PATHS
# ══════════════════════════════════════════════════════════
paths:
...
  /api/v1/auth/signin:
    post:
      tags: [auth]
      operationId: signIn
      summary: Sign in or create account
      description: |
        Authenticates with Apple, Google, or email/password. Creates an account
        on first sign-in (no separate registration endpoint needed).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/SignInApple'
                - $ref: '#/components/schemas/SignInGoogle'
                - $ref: '#/components/schemas/SignInEmail'
              discriminator:
                propertyName: provider
                mapping:
                  apple:  '#/components/schemas/SignInApple'
                  google: '#/components/schemas/SignInGoogle'
                  email:  '#/components/schemas/SignInEmail'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      description: >-
        Rotates the refresh token: the presented token is invalidated and a new
        access + refresh pair is returned. Clients must persist the returned
        refresh_token. Replaying an already-rotated token is treated as reuse and
        revokes every session for the user.
      responses:
        '200':
          description: New access + refresh token pair issued
          content:
            application/json:
              schema:
                type: object
                required: [access_token, refresh_token, expires_in]
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                    description: Rotated refresh token — replaces the one sent.
                  expires_in:
                    type: integer
                    example: 3600
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Invalidate session
      description: Revokes the refresh token. The short-lived access token expires naturally.
      responses:
        '204':
          description: Session invalidated
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/verify-phone:
    post:
      tags: [auth]
      operationId: verifyPhone
      summary: Verify phone number via SMS OTP
      description: |
        Two-step flow:
        1. Call without `otp` → server sends a 6-digit SMS to `phone_number`
        2. Call with `otp` → server verifies; success raises the reputation base score

        Phone verification is optional but strongly recommended.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number]
              properties:
                phone_number:
                  type: string
                  example: "+4915112345678"
                otp:
                  type: string
                  pattern: '^\d{6}$'
                  description: Omit to request a new code; include to verify
      responses:
        '200':
          description: Code sent or verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/device-token:
    put:
      tags: [auth]
      operationId: updateDeviceToken
      summary: Register or update APNs device token
      description: |
        Must be called after every sign-in and whenever iOS assigns a new push token.
        Used for alarm notifications and silent wakeups (Dead Man's Switch, v2.0).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_token, environment]
              properties:
                device_token:
                  type: string
                  description: Hex-encoded APNs device token from iOS
                environment:
                  type: string
                  enum: [sandbox, production]
      responses:
        '204':
          description: Device token stored
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Users ─────────────────────────────────────────────────
...